+++ src/lib/components/table/proposals/ProposalTextCell.tsx
@@ -1,6 +1,7 @@
 import { Flex, Text } from "@chakra-ui/react";
 import { useRef, useState } from "react";
 
+import { MobileLabel } from "../MobileLabel";
 import { useMobile } from "lib/app-provider";
 import { DotSeparator } from "lib/components/DotSeparator";
 import { Expedited } from "lib/components/Expedited";
@@ -31,13 +32,51 @@ export const ProposalTextCell = ({
       Number(typeRef.current?.scrollWidth) >
         Number(typeRef.current?.clientWidth));
 
+  if (isMobile) {
+    return (
+      <Flex direction="column" gap={1}>
+        <MobileLabel label="Proposal Title" />
+        <Text color="text.main" variant="body2" wordBreak="break-word">
+          {title}
+          {isExpedited && (
+            <span
+              style={{
+                display: "inline-block",
+                marginLeft: "8px",
+                verticalAlign: "middle",
+              }}
+            >
+              <Expedited isActiveExpedited={isDepositOrVoting} />
+            </span>
+          )}
+        </Text>
+        {types.length ? (
+          types.map((msgType, index) => (
+            <Text
+              key={msgType + index.toString()}
+              variant="body3"
+              color="text.dark"
+              wordBreak="break-word"
+            >
+              {msgType}
+            </Text>
+          ))
+        ) : (
+          <Text variant="body3" color="text.dark">
+            (No message)
+          </Text>
+        )}
+      </Flex>
+    );
+  }
+
   return (
     <Flex
       flexDirection="column"
       justify="center"
       borderRadius="8px"
       bgColor={showName ? "gray.800" : "undefined"}
-      px={isMobile ? 0 : 4}
+      px={4}
       maxW={showName ? undefined : "full"}
       onMouseOver={() => setIsHoverText(true)}
       onMouseOut={() => setIsHoverText(false)}
@@ -74,7 +113,7 @@ export const ProposalTextCell = ({
                 <span key={msgType + index.toString()}>
                   {index > 0 && (
                     <span style={{ color: "var(--chakra-colors-accent-main)" }}>
-                      {" / "}
+                      {" , "}
                     </span>
                   )}
                   {msgType}