+++ src/lib/pages/validator-details/components/tables/voted-proposals/index.tsx
@@ -2,15 +2,20 @@ import { Alert, Flex, Grid, GridItem, Text } from "@chakra-ui/react";
 import type { ChangeEvent } from "react";
 import { useMemo, useState } from "react";
 
+import { AmpEvent, trackUseFilter, trackUseViewMore } from "lib/amplitude";
 import { useMobile } from "lib/app-provider";
 import { SelectInput } from "lib/components/forms";
+import type { IconKeys } from "lib/components/icon";
 import { CustomIcon } from "lib/components/icon";
 import InputWithIcon from "lib/components/InputWithIcon";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { TableTitle, ViewMore } from "lib/components/table";
 import { useDebounce } from "lib/hooks";
-import { useValidatorVotedProposals } from "lib/services/validatorService";
+import {
+  useValidatorVotedProposals,
+  useValidatorVotedProposalsAnswerCounts,
+} from "lib/services/validatorService";
 import { ProposalVoteType } from "lib/types";
 import type { ValidatorAddr } from "lib/types";
 
@@ -59,46 +64,61 @@ export const VotedProposalsTable = ({
     debouncedSearch,
     { onSuccess: ({ total }) => setTotalData(total) }
   );
+  const { data: answers } =
+    useValidatorVotedProposalsAnswerCounts(validatorAddress);
 
   const answerOptions = useMemo(
     () => [
       {
-        label: `All votes`,
+        label: `All proposals (${answers?.all ?? 0})`,
         value: ProposalVoteType.ALL,
-        disabled: false,
+        disabled: !answers?.all,
       },
       {
-        label: `Yes`,
+        label: `Yes (${answers?.yes ?? 0})`,
         value: ProposalVoteType.YES,
-        disabled: false,
+        disabled: !answers?.yes,
+        icon: "circle" as IconKeys,
+        iconColor: "success.main",
       },
       {
-        label: `No`,
+        label: `No (${answers?.no ?? 0})`,
         value: ProposalVoteType.NO,
-        disabled: false,
+        disabled: !answers?.no,
+        icon: "circle" as IconKeys,
+        iconColor: "error.main",
       },
       {
-        label: `No with veto`,
+        label: `No with veto (${answers?.noWithVeto ?? 0})`,
         value: ProposalVoteType.NO_WITH_VETO,
-        disabled: false,
+        disabled: !answers?.noWithVeto,
+        icon: "circle" as IconKeys,
+        iconColor: "error.dark",
       },
       {
-        label: `Abstain`,
+        label: `Abstain (${answers?.abstain ?? 0})`,
         value: ProposalVoteType.ABSTAIN,
-        disabled: false,
+        disabled: !answers?.abstain,
+        icon: "circle" as IconKeys,
+        iconColor: "gray.600",
       },
       {
-        label: `Weighted`,
-        value: ProposalVoteType.WEIGHTED,
-        disabled: false,
-      },
-      {
-        label: `Did not vote`,
+        label: `Did not vote (${answers?.didNotVote ?? 0})`,
         value: ProposalVoteType.DID_NOT_VOTE,
-        disabled: false,
+        disabled: !answers?.didNotVote,
+        icon: "circle" as IconKeys,
+        iconColor: "gray.600",
+      },
+      {
+        label: `Weighted (${answers?.weighted ?? 0})`,
+        value: ProposalVoteType.WEIGHTED,
+        disabled: !answers?.weighted,
+        icon: "circle" as IconKeys,
+        iconColor: "primary.light",
       },
     ],
-    []
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [JSON.stringify(answers)]
   );
 
   const handleOnSearchChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -107,6 +127,11 @@ export const VotedProposalsTable = ({
   };
 
   const handleOnAnswerFilterChange = (newAnswer: ProposalVoteType) => {
+    trackUseFilter(
+      AmpEvent.USE_FILTER_VOTED_PROPOSALS_ANSWER,
+      Object.values(ProposalVoteType),
+      newAnswer
+    );
     setCurrentPage(1);
     setAnswerFilter(newAnswer);
   };
@@ -119,14 +144,17 @@ export const VotedProposalsTable = ({
       w="100%"
       justifyContent="space-between"
       alignItems="center"
-      onClick={onViewMore}
+      onClick={() => {
+        trackUseViewMore();
+        onViewMore();
+      }}
     >
-      <TableTitle title="Voted Proposals" count={data?.total ?? 0} mb={0} />
+      <TableTitle title="Voted Proposals" count={answers?.all ?? 0} mb={0} />
       <CustomIcon boxSize={6} m={0} name="chevron-right" color="gray.600" />
     </Flex>
   ) : (
     <Flex direction="column" gap={6}>
-      <TableTitle title="Voted Proposals" count={data?.total ?? 0} mb={0} />
+      <TableTitle title="Voted Proposals" count={answers?.all ?? 0} mb={0} />
       {!onViewMore && (
         <>
           <Alert variant="info" gap={4} display={{ base: "none", md: "flex" }}>
@@ -154,6 +182,7 @@ export const VotedProposalsTable = ({
                 value={search}
                 onChange={handleOnSearchChange}
                 size="lg"
+                amptrackSection="voted-proposals-search"
               />
             </GridItem>
           </Grid>
@@ -163,12 +192,13 @@ export const VotedProposalsTable = ({
         data={data}
         isLoading={isLoading}
         onViewMore={onViewMore}
+        search={search}
       />
       {data &&
         (onViewMore
           ? data.total > 5 && (
               <ViewMore
-                text={`View all proposed blocks (${data.total})`}
+                text={`View all proposals (${data.total})`}
                 onClick={onViewMore}
               />
             )